examples: usage examples
Add a collection of usage examples.

Covers all UI elements. Future PRs will add examples for other things
our users may want to get done.
