+++ examples/README.md
@@ -1,10 +1,14 @@
 # Examples
 
-This folder contains example marimo notebooks.
+This folder contains example marimo notebooks, meant to teach you how to
+use marimo's features as well as inspire you to make awesome notebooks.
 
-- 🖱️ [`ui/`](ui/): marimo's many UI elements or widgets
-- 🛢️ [`sql/`](sql/): using marimo's SQL cells
-- 📽️ [`layouts/`](layouts/): present notebooks as slides, add sidebars, and more
+
+- 🖱️ [`ui/`](ui/): how to use UI elements or widgets
+- 🛢️ [`sql/`](sql/): how to use SQL in marimo
+- 🛢️ [`control_flow/`](control_flow/): how to control cell execution and output display
+- 🛢️ [`markdown/`](markdown/): how to write markdown, including dynamic markdown
+- 📽️ [`layouts/`](layouts/): how to present notebooks as slides, add sidebars, and more
 - 🤖 [`ai/`](ai/): AI-related examples
 - 📦 [`third_party/`](third_party/): using popular third-party packages in marimo
 - ☁️  [`cloud/`](cloud/): using various cloud providers
@@ -12,13 +16,16 @@ This folder contains example marimo notebooks.
 - ✨ [`misc/`](misc/): miscellaneous topical examples
 
 > [!TIP]
-> New to marimo? Run `marimo tutorial ui` at the command line first!
+> New to marimo? Run `marimo tutorial intro` at the command line first!
+
+> [!TIP]
+> Check out our [public gallery](https://marimo.io/gallery) of interactive
+> notebooks to get inspired.
 
 > [!NOTE]
 > Submit a
 > [pull request](https://github.com/marimo-team/marimo/pulls) to add an example!
 > We especially welcome library developers to add examples to `third_party/`.
-
 
 ## Running examples
