+++ examples/control_flow/stop_execution.py
