--- examples/getting_started/inputs/refresh.py
@@ -1,27 +1,0 @@
-import marimo
-
-__generated_with = "0.10.6"
-app = marimo.App()
-
-
-@app.cell
-def _():
-    import marimo as mo
-    return (mo,)
-
-
-@app.cell
-def _(mo):
-    refresh = mo.ui.refresh(default_interval=1)
-    refresh
-    return (refresh,)
-
-
-@app.cell
-def _(refresh):
-    print(refresh.value)
-    return
-
-
-if __name__ == "__main__":
-    app.run()
