+++ examples/markdown/mermaid.py
