+++ examples/misc/reactive_plots.py
