+++ examples/ui/batch.py
