+++ examples/ui/slider.py
