+++ examples/ui/switch.py
