+++ examples/ui/tabs.py
@@ -1,74 +1,29 @@
-# /// script
-# requires-python = ">=3.9"
-# dependencies = [
-#     "marimo",
-# ]
-# ///
-
 import marimo
 
-__generated_with = "0.8.19"
+__generated_with = "0.10.6"
 app = marimo.App()
 
 
-@app.cell(hide_code=True)
-def __(mo):
-    mo.md("""# Tabs""")
-    return
-
-
-@app.cell(hide_code=True)
-def __(mo):
-    mo.md("""Use `mo.ui.tabs` to organize outputs.""")
-    return
-
-
 @app.cell
-def __(mo):
-    settings = mo.vstack(
-        [
-            mo.md("Edit User"),
-            first := mo.ui.text(label="First Name"),
-            last := mo.ui.text(label="Last Name"),
-        ]
-    )
-
-    organization = mo.vstack(
-        [
-            mo.md("Edit Organization"),
-            org := mo.ui.text(label="Organization Name", value="..."),
-            employees := mo.ui.number(
-                label="Number of Employees", start=0, stop=1000
-            ),
-        ]
-    )
-
-    mo.ui.tabs(
-        {
-            "🧙‍♀ User": settings,
-            "🏢 Organization": organization,
-        }
-    )
-    return employees, first, last, org, organization, settings
-
-
-@app.cell
-def __(employees, first, last, mo, org):
-    mo.md(
-        f"""
-        Welcome **{first.value} {last.value}** to **{org.value}**! You are 
-        employee no. **{employees.value + 1}**.
-
-        #{"🎉" * (min(employees.value + 1, 1000))} 
-        """
-    ) if all([first.value, last.value, org.value]) else None
-    return
-
-
-@app.cell
-def __():
+def _():
     import marimo as mo
     return (mo,)
+
+
+@app.cell
+def _(mo):
+    tabs = mo.ui.tabs({
+        "Bob says": mo.md("Hello, Alice! 👋"),
+        "Alice says": mo.md("Hello, Bob! 👋")
+    })
+    tabs
+    return (tabs,)
+
+
+@app.cell
+def _(tabs):
+    tabs.value
+    return
 
 
 if __name__ == "__main__":
