+++ examples/ui/text.py
